+++ src/components/Dashboard/DashboardItems/ItemAmount.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { Torrent } from '@/types/VueTorrent'
+import { Torrent } from '@/types/vuetorrent'
 
 defineProps<{ torrent: Torrent, title: string, value: string, total: string }>()
 </script>
