+++ src/components/Dashboard/DashboardItems/ItemDuration.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { Torrent } from '@/types/VueTorrent'
+import { Torrent } from '@/types/vuetorrent'
 import dayjs from '@/plugins/dayjs'
 
 defineProps<{ torrent: Torrent, title: string, value: string }>()
