+++ src/components/Dashboard/DashboardItems/ItemPercent.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { TorrentState } from '@/constants/qbit'
 import { formatPercent } from '@/helpers'
-import { Torrent } from '@/types/VueTorrent'
+import { Torrent } from '@/types/vuetorrent'
 import { computed } from 'vue'
 
 const props = defineProps<{ torrent: Torrent, title: string, value: string }>()
