+++ src/components/Dashboard/DashboardItems/ItemText.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { Torrent } from '@/types/VueTorrent'
+import { Torrent } from '@/types/vuetorrent'
 
 defineProps<{ torrent: Torrent, title: string, value: string }>()
 </script>
