+++ src/components/TorrentDetail/Content.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { Torrent } from '@/types/VueTorrent'
+import { Torrent } from '@/types/vuetorrent'
 
 const props = defineProps<{ torrent: Torrent, isActive: boolean }>()
 </script>
