+++ src/constants/vuetorrent/index.ts
@@ -1,8 +1,11 @@
+import { propsData, propsMetadata } from './DashboardDefaults'
 import { DashboardProperty } from './DashboardProperty'
 import { DashboardPropertyType } from './DashboardPropertyType'
 import { TitleOptions, titleOptionsList } from './TitleOptions'
 
 export {
+  propsData,
+  propsMetadata,
   DashboardProperty,
   DashboardPropertyType,
   TitleOptions,
