+++ src/pages/Dashboard.vue
@@ -4,7 +4,7 @@ import RightClickMenu from '@/components/Dashboard/TRC/RightClickMenu.vue'
 import ConfirmDeleteDialog from '@/components/Dialogs/ConfirmDeleteDialog.vue'
 import { doesCommand } from '@/helpers'
 import Torrent from '@/components/Dashboard/Torrent.vue'
-import { Torrent as TorrentType } from '@/types/VueTorrent'
+import { Torrent as TorrentType } from '@/types/vuetorrent'
 import { useDashboardStore, useMaindataStore, useNavbarStore, useVueTorrentStore } from '@/stores'
 import { computed, nextTick, onBeforeMount, onBeforeUnmount, onMounted, reactive, ref, watch } from 'vue'
 import { useI18n } from 'vue-i18n'
