+++ src/pages/RssArticles.vue
@@ -3,7 +3,7 @@ import debounce from 'lodash.debounce'
 import { useSearchQuery } from '@/composables'
 import { useNavbarStore, useRssStore } from '@/stores'
 import { Feed } from '@/types/qbit/models'
-import { RssArticle } from '@/types/VueTorrent'
+import { RssArticle } from '@/types/vuetorrent'
 import { computed, onBeforeMount, onMounted, onUnmounted } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { useRouter } from 'vue-router'
