+++ src/pages/SearchEngine.vue
@@ -3,7 +3,7 @@ import PluginManagerDialog from '@/components/Dialogs/PluginManagerDialog.vue'
 import { useSearchQuery } from '@/composables'
 import { useNavbarStore, useSearchEngineStore, useVueTorrentStore } from '@/stores'
 import { SearchPlugin, SearchResult } from '@/types/qbit/models'
-import { SearchData } from '@/types/VueTorrent'
+import { SearchData } from '@/types/vuetorrent'
 import { storeToRefs } from 'pinia'
 import { formatData } from '@/helpers'
 import { computed, onBeforeMount, onBeforeUnmount, ref } from 'vue'
