+++ src/types/vuetorrent/RssArticle.ts
