+++ src/types/vuetorrent/SearchData.ts
