+++ src/types/vuetorrent/TRCMenuEntry.ts
