+++ src/types/vuetorrent/Torrent.ts
