+++ src/types/vuetorrent/TorrentProperty.ts
